refactor: rename substrate endpoint folder.
