+++ src/app.module.ts
@@ -6,7 +6,7 @@ import { LocationModule } from './endpoints/location/location.module';
 import { RatingModule } from './endpoints/rating/rating.module';
 import { EthereumModule } from './endpoints/ethereum/ethereum.module';
 import { EscrowModule } from './endpoints/escrow/escrow.module';
-import { SubstrateEndpointModule } from './endpoints/substrate/substrate-endpoint.module';
+import { SubstrateEndpointModule } from './endpoints/substrate-endpoint/substrate-endpoint.module';
 import { EthereumIndexedDataModule } from './endpoints/ethereum-indexed-data/ethereum-indexed-data.module';
 import { TransactionLoggingModule } from './common/modules/transaction-logging/transaction-logging.module';
 import { TransactionRequest } from './common/modules/transaction-logging/models/transaction-request.entity';
