+++ src/endpoints/substrate-endpoint/dto/get-dbio-on-register.dto.ts
