+++ src/endpoints/substrate-endpoint/dto/wallet-binding.dto.ts
