+++ src/endpoints/substrate-endpoint/labs/lab.controller.ts
