+++ test/unit/endpoints/substrate-indexed-data/labs/lab.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { when } from 'jest-when';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
-import { LabService } from '../../../../../src/endpoints/substrate/services/lab.service';
-import { LabController } from '../../../../../src/endpoints/substrate/labs/lab.controller';
+import { LabService } from '../../../../../src/endpoints/substrate-endpoint/services/lab.service';
+import { LabController } from '../../../../../src/endpoints/substrate-endpoint/labs/lab.controller';
 import { elasticsearchServiceMockFactory, MockType } from '../../../mock';
 
 describe('Substrate Indexer Lab Controller Unit Tests', () => {
