+++ test/unit/endpoints/substrate-indexed-data/services/service.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { when } from 'jest-when';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
-import { ServiceService } from '../../../../../src/endpoints/substrate/services/service.service';
+import { ServiceService } from '../../../../../src/endpoints/substrate-endpoint/services/service.service';
 import { elasticsearchServiceMockFactory, MockType } from '../../../mock';
 
 describe('Substrate Indexer Service Service Unit Tests', () => {
